+++ src/components/forms/Form.ts
@@ -31,6 +31,8 @@ export type FormOptions =
 
 	manuallyInitialize?: boolean;
 	maxFileSize?: number;
+
+	/** @deprecated */
 	noticeContainerSelector?: string;
 
 	protection?: FormProtectionOptions;
@@ -50,18 +52,19 @@ export function Form(options: FormOptions, children: Child)
 	}
 
 	const method = options.method ?? "get";
-
 	const action = options.action ?? "";
 
 	const autoComplete = options.autoComplete ?? true;
-
 	const encodingType = options.encodingType ?? "application/x-www-form-urlencoded";
 
 	const manuallyInitialize = options.manuallyInitialize ?? false;
-
 	const maxFileSize = options.maxFileSize ?? -1;
 
 	const protection = options.protection ?? { type: "none" };
+
+	const hiddenInputs = options.hiddenInputs ?? {};
+
+	const extraAttributes = options.extraAttributes ?? {};
 
 	const protectionAttributes: Record<string, string> = {};
 
@@ -83,8 +86,6 @@ export function Form(options: FormOptions, children: Child)
 		}
 	}
 
-	const hiddenInputs = options.hiddenInputs ?? {};
-
 	return new DE("form",
 		{
 			"class": className,
@@ -96,14 +97,13 @@ export function Form(options: FormOptions, children: Child)
 
 			"data-manually-initialize": manuallyInitialize,
 			"data-max-file-size": maxFileSize,
-			"data-notice-container-selector": options.noticeContainerSelector ?? ".notice-container",
 
 			...protectionAttributes,
 
-			...options.extraAttributes,
+			...extraAttributes,
 		},
 		[
-			new DE("div", "hidden",
+			new DE("div", "hidden-container",
 				[
 					Object.entries(hiddenInputs).map(
 						([ name, value ]) =>
@@ -112,6 +112,8 @@ export function Form(options: FormOptions, children: Child)
 						}),
 				]),
 
-			new DE("div", "inputs", children),
+			new DE("div", "notice-container"),
+
+			new DE("div", "input-container", children),
 		]);
 }